+++ handlers/index.js
@@ -1,0 +1,30 @@
+const handleWelcome = require('./welcome');
+const handleBan = require('./ban');
+const handleUnban = require('./unban');
+const handleHelp = require('./help');
+const handleF = require('./f');
+const handleTest = require('./test');
+const handleGifID = require('./gifid');
+const handleKick = require('./kick');
+const handleShrug = require('./shrug');
+const handleReg = require('./reg');
+const handleChecker = require('./checker');
+const handleSetRules = require('./setrules');
+const handleRules = require('./rules');
+const handlePidor = require('./pidor');
+module.exports = {
+  handleWelcome,
+  handleBan,
+  handleUnban,
+  handleHelp,
+  handleF,
+  handleTest,
+  handleGifID,
+  handleKick,
+  handleReg,
+  handleShrug,
+  handleChecker,
+  handleSetRules,
+  handleRules,
+  handlePidor,
+};