+++ CHANGELOG.md
@@ -1,34 +1,78 @@
 # Changelog
 
-## v0.1.0
+## v.1.0.0
 
-- Commands `/ban`, `/kick`, `/unban`
-- Added welcome message to new chat members
+- Updated code base
+
+- Bot now replies to messages
+
+- New commands: `/rules`, `/setrules`
+
+- Bot now sends number of version on `/test`
+
+- Bot now checks new chat member in the database
+
+- Changelog in reverse order
+
+- Readme in English
+
+- Added catch blocks
+
+- added config/ to .gitignore
+
+- updated Mongoose to 5.8.9
+
+## v.0.4.1
+
+- Checking the user for the presence in the database
+
+## v.0.4.0
+
+- Added support of MongoDB
+
+## v.0.3.6
+
+- Added output of version on /test
+
+## v.0.3.5
+
+- Added command `/shrug`
+- gifs.json moved to config directory
+
+## v.0.3.4
+
+- Added command `/f`
+- Added trigger **gifid** for getting file_id of the GIF
+
+## v.0.3.3
+
 - Bug fixes
 
-## v0.2.0
+## v.0.3.2
 
-- Bug fixes
-- Updated dependencies
-- Added Eslint to devDependencies
-- Updated README
+- Added support of Markdown in replies
 
-## v0.3.0
+## v.0.3.1
+
+- Updated Telegraf to 3.35.0
+- Bot now sends errors to creator in Telegram
+
+## v.0.3.0
 
 - Bug fixes
 - Updated dependencies
 - Replies moved to yaml
 - Support multilanguage
 
-## v0.3.1
+## v.0.2.0
 
-- Updated Telegraf to 3.35.0
-- Bot now sends errors to creator in Telegram
+- Bug fixes
+- Updated dependencies
+- Added Eslint to devDependencies
+- Updated README
 
-## v.0.3.2
+## v.0.1.0
 
-- Added support of Markdown in replies
-
-## v.0.3.3
-
+- Commands `/ban`, `/kick`, `/unban`
+- Added welcome message to new chat members
 - Bug fixes